# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/service": {
      "post": {
        "description": "Permite a un administrador crear un nuevo servicio médico en el sistema (Ej. \"Consultoría Nutricional\", \"Electrocardiograma\").\n\n**Nota**: Se requiere enviar un `areaId` válido que relacione este servicio con una especialidad/área del centro médico.",
        "operationId": "ServiceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Servicio creado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "12345678-1234-1234-1234-123456789012",
                    "name": "Consulta de control",
                    "description": "Chequeo de rutina",
                    "price": 100,
                    "duration": 60,
                    "areaId": "12345678-1234-1234-1234-123456789012",
                    "isActive": true,
                    "createdAt": "2022-01-01T00:00:00.000Z",
                    "updatedAt": "2022-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un nuevo servicio",
        "tags": [
          "Service"
        ]
      },
      "get": {
        "description": "Obtiene el catálogo de servicios ofrecidos. \n    \nÚtil para mostrarle al Doctor o Administrador las opciones disponibles al momento de **crear una consulta**.\nEl arreglo devuelto contendrá el `id` que se debe usar en el campo `serviceIds` de la consulta.",
        "operationId": "ServiceController_findAll",
        "parameters": [
          {
            "name": "isActive",
            "required": true,
            "in": "path",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Servicios obtenidos exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "12345678-1234-1234-1234-123456789012",
                      "name": "Consultoría Nutricional",
                      "description": "Evaluación y orientación nutricional personalizada.",
                      "price": 100,
                      "duration": 60,
                      "areaId": "12345678-1234-1234-1234-123456789012",
                      "isActive": true,
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "updatedAt": "2022-01-01T00:00:00.000Z",
                      "area": {
                        "id": "12345678-1234-1234-1234-123456789012",
                        "name": "Nutrición",
                        "description": "Área encargada de la nutrición",
                        "isActive": true,
                        "createdAt": "2022-01-01T00:00:00.000Z",
                        "updatedAt": "2022-01-01T00:00:00.000Z"
                      },
                      "consultations": []
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Obtener todos los servicios",
        "tags": [
          "Service"
        ]
      }
    },
    "/service/{id}": {
      "get": {
        "operationId": "ServiceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Servicio obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "12345678-1234-1234-1234-123456789012",
                    "name": "Consultoría Nutricional",
                    "description": "Evaluación y orientación nutricional personalizada.",
                    "price": 100,
                    "duration": 60,
                    "areaId": "12345678-1234-1234-1234-123456789012",
                    "isActive": true,
                    "createdAt": "2022-01-01T00:00:00.000Z",
                    "updatedAt": "2022-01-01T00:00:00.000Z",
                    "area": {
                      "id": "12345678-1234-1234-1234-123456789012",
                      "name": "Nutrición",
                      "description": "Área encargada de la nutrición",
                      "isActive": true,
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "updatedAt": "2022-01-01T00:00:00.000Z"
                    },
                    "consultations": []
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener un servicio por ID",
        "tags": [
          "Service"
        ]
      },
      "patch": {
        "operationId": "ServiceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Servicio actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "12345678-1234-1234-1234-123456789012",
                    "name": "Consultoría Nutricional",
                    "description": "Evaluación y orientación nutricional personalizada.",
                    "price": 100,
                    "duration": 60,
                    "areaId": "12345678-1234-1234-1234-123456789012",
                    "isActive": true,
                    "createdAt": "2022-01-01T00:00:00.000Z",
                    "updatedAt": "2022-01-01T00:00:00.000Z",
                    "area": {
                      "id": "12345678-1234-1234-1234-123456789012",
                      "name": "Nutrición",
                      "description": "Área encargada de la nutrición",
                      "isActive": true,
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "updatedAt": "2022-01-01T00:00:00.000Z"
                    },
                    "consultations": []
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un servicio",
        "tags": [
          "Service"
        ]
      }
    },
    "/service/{id}/disable": {
      "patch": {
        "operationId": "ServiceController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Servicio desactivado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "12345678-1234-1234-1234-123456789012",
                    "name": "Consultoría Nutricional",
                    "description": "Evaluación y orientación nutricional personalizada.",
                    "price": 100,
                    "duration": 60,
                    "areaId": "12345678-1234-1234-1234-123456789012",
                    "isActive": false,
                    "createdAt": "2022-01-01T00:00:00.000Z",
                    "updatedAt": "2022-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar un servicio",
        "tags": [
          "Service"
        ]
      }
    },
...
  },
  "info": {
    "title": "API CLIMAS",
    "description": "Documentación API CLIMAS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del servicio",
            "example": "Consulta Médica"
          },
          "description": {
            "type": "string",
            "description": "Descripción del servicio",
            "example": "Evaluación médica general por un profesional"
          },
          "price": {
            "type": "number",
            "minimum": 1,
            "description": "Precio del servicio",
            "example": 50000
          },
          "duration": {
            "type": "number",
            "minimum": 1,
            "description": "Duración del servicio en minutos",
            "example": 30
          },
          "areaId": {
            "type": "string",
            "description": "ID del área a la que pertenece el servicio",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indica si el servicio está activo",
            "example": true
          }
        },
        "required": [
          "name",
          "areaId"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del servicio",
            "example": "Consulta Médica"
          },
          "description": {
            "type": "string",
            "description": "Descripción del servicio",
            "example": "Evaluación médica general por un profesional"
          },
          "price": {
            "type": "number",
            "minimum": 1,
            "description": "Precio del servicio",
            "example": 50000
          },
          "duration": {
            "type": "number",
            "minimum": 1,
            "description": "Duración del servicio en minutos",
            "example": 30
          },
          "areaId": {
            "type": "string",
            "description": "ID del área a la que pertenece el servicio",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indica si el servicio está activo",
            "example": true
          }
        }
      },
...
    }
  }
}